requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.5
pillow==11.0.0
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
torch==2.7.0
torchvision==0.22.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
